Kiosk/MicroControllers/heartbeat_service/requirements.txt
max30102-python
python-socketio[client]
requests
